api/requirements.txt
Flask>=1.1.1
Flask-Cors>=3.0.8
Flask-RESTful>=0.3.7
Werkzeug==0.15.6
cassandra-driver>=3.19.0